+++ backend/routes/users.ts
@@ -17,35 +17,39 @@ users.get("/users", async (_, res, next) => {
     },
   });
   res.status(200).json(users);
-  console.log("get 200");
   next();
 });
 
 users.patch("/users", async (req, res, next) => {
-  const usernames = req.body.data.usernames;
-  const active = req.body.data.active;
+  const usernames = req.body.usernames;
+  const active = req.body.active;
+  if (!usernames || !active) {
+    res.status(422).send();
+    return;
+  }
   await prisma.user.updateMany({
     where: { username: { in: usernames } },
     data: { active },
   });
   res.status(204);
-  console.log("patch 204");
   next();
 });
 
 users.delete("/users", async (req, res, next) => {
   const usernames = req.body.usernames;
+  if (!usernames) {
+    res.status(422).send();
+    return;
+  }
   await prisma.user.deleteMany({
     where: { username: { in: usernames } },
   });
   res.status(204);
-  console.log("delete 204");
   next();
 });
 
 users.use(checkActiveStatus);
 users.use(async (_, res) => {
-  console.log("at this point everythin is good");
   res.send();
 });
